Fixed Requirements Bug - Part I
So, when creating a new requirement, while all of the choices were
available for the RAM field, it would not allow you to save it, citing
an error on the field saying that '1' or '2' or '4' or any of the
options were available choices.  So I'm not sure what I did wrong.  The
way I fixed it was to change the tuple options to mimic each other: ('1
GB', '1 GB') instead of ('1', '1 GB'), and I changed the field type to
CHAR instead of INT.  This will need future work.